appium/python/pytest/requirements.txt
Appium-Python-Client==2.3.0;python_version >= '3.0'
selenium==4.3.0
pytest-dogu-report==1.0.1
pytest==7.4.0
dogu-device-client==0.0.5